Changed classes to have more clear purpose
